+++ fractal_server/app/routes/auth/users.py
@@ -8,9 +8,7 @@ from fastapi import status
 from fastapi_users import exceptions
 from fastapi_users import schemas
 from fastapi_users.router.common import ErrorCode
-from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlmodel import col
 from sqlmodel import func
 from sqlmodel import select
 
@@ -18,9 +16,10 @@ from . import current_active_superuser
 from ...db import get_async_db
 from ...schemas.user import UserRead
 from ...schemas.user import UserUpdate
-from ...schemas.user import UserUpdateWithNewGroupIds
 from ..aux.validate_user_settings import verify_user_has_settings
+from ._aux_auth import _get_default_usergroup_id
 from ._aux_auth import _get_single_user_with_groups
+from ._aux_auth import FRACTAL_DEFAULT_GROUP_NAME
 from fractal_server.app.models import LinkUserGroup
 from fractal_server.app.models import UserGroup
 from fractal_server.app.models import UserOAuth
@@ -28,6 +27,7 @@ from fractal_server.app.models import UserSettings
 from fractal_server.app.routes.auth._aux_auth import _user_or_404
 from fractal_server.app.schemas import UserSettingsRead
 from fractal_server.app.schemas import UserSettingsUpdate
+from fractal_server.app.schemas.user import UserUpdateGroups
 from fractal_server.app.security import get_user_manager
 from fractal_server.app.security import UserManager
 from fractal_server.logger import set_logger
@@ -55,114 +55,43 @@ async def get_user(
 @router_users.patch("/users/{user_id}/", response_model=UserRead)
 async def patch_user(
     user_id: int,
-    user_update: UserUpdateWithNewGroupIds,
+    user_update: UserUpdate,
     current_superuser: UserOAuth = Depends(current_active_superuser),
     user_manager: UserManager = Depends(get_user_manager),
     db: AsyncSession = Depends(get_async_db),
 ):
     """
     Custom version of the PATCH-user route from `fastapi-users`.
-
-    In order to keep the fastapi-users logic in place (which is convenient to
-    update user attributes), we split the endpoint into two branches. We either
-    go through the fastapi-users-based attribute-update branch, or through the
-    branch where we establish new user/group relationships.
-
-    Note that we prevent making both changes at the same time, since it would
-    be more complex to guarantee that endpoint error would leave the database
-    in the same state as before the API call.
     """
-
-    # We prevent simultaneous editing of both user attributes and user/group
-    # associations
-    user_update_dict_without_groups = user_update.dict(
-        exclude_unset=True, exclude={"new_group_ids"}
-    )
-    edit_attributes = user_update_dict_without_groups != {}
-    edit_groups = user_update.new_group_ids is not None
-    if edit_attributes and edit_groups:
-        raise HTTPException(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail=(
-                "Cannot modify both user attributes and group membership. "
-                "Please make two independent PATCH calls"
-            ),
-        )
 
     # Check that user exists
     user_to_patch = await _user_or_404(user_id, db)
 
-    if edit_groups:
-        # Establish new user/group relationships
-
-        # Check that all required groups exist
-        # Note: The reason for introducing `col` is as in
-        # https://sqlmodel.tiangolo.com/tutorial/where/#type-annotations-and-errors,
-        stm = select(func.count()).where(
-            col(UserGroup.id).in_(user_update.new_group_ids)
+    # Modify user attributes
+    try:
+        user = await user_manager.update(
+            user_update,
+            user_to_patch,
+            safe=False,
+            request=None,
         )
-        res = await db.execute(stm)
-        number_matching_groups = res.scalar()
-        if number_matching_groups != len(user_update.new_group_ids):
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=(
-                    "Not all requested groups (IDs: "
-                    f"{user_update.new_group_ids}) exist."
-                ),
-            )
-
-        for new_group_id in user_update.new_group_ids:
-            link = LinkUserGroup(user_id=user_id, group_id=new_group_id)
-            db.add(link)
-
-        try:
-            await db.commit()
-        except IntegrityError as e:
-            error_msg = (
-                f"Cannot link groups with IDs {user_update.new_group_ids} "
-                f"to user {user_id}. "
-                "Likely reason: one of these links already exists.\n"
-                f"Original error: {str(e)}"
-            )
-            logger.info(error_msg)
-            raise HTTPException(
-                status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-                detail=error_msg,
-            )
-        patched_user = user_to_patch
-    elif edit_attributes:
-        # Modify user attributes
-        try:
-            user_update_without_groups = UserUpdate(
-                **user_update_dict_without_groups
-            )
-            user = await user_manager.update(
-                user_update_without_groups,
-                user_to_patch,
-                safe=False,
-                request=None,
-            )
-            validated_user = schemas.model_validate(UserOAuth, user)
-            patched_user = await db.get(
-                UserOAuth, validated_user.id, populate_existing=True
-            )
-        except exceptions.InvalidPasswordException as e:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail={
-                    "code": ErrorCode.UPDATE_USER_INVALID_PASSWORD,
-                    "reason": e.reason,
-                },
-            )
-        except exceptions.UserAlreadyExists:
-            raise HTTPException(
-                status.HTTP_400_BAD_REQUEST,
-                detail=ErrorCode.UPDATE_USER_EMAIL_ALREADY_EXISTS,
-            )
-    else:
-        # Nothing to do, just continue
-        patched_user = user_to_patch
+        validated_user = schemas.model_validate(UserOAuth, user)
+        patched_user = await db.get(
+            UserOAuth, validated_user.id, populate_existing=True
+        )
+    except exceptions.InvalidPasswordException as e:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail={
+                "code": ErrorCode.UPDATE_USER_INVALID_PASSWORD,
+                "reason": e.reason,
+            },
+        )
+    except exceptions.UserAlreadyExists:
+        raise HTTPException(
+            status.HTTP_400_BAD_REQUEST,
+            detail=ErrorCode.UPDATE_USER_EMAIL_ALREADY_EXISTS,
+        )
 
     # Enrich user object with `group_ids_names` attribute
     patched_user_with_groups = await _get_single_user_with_groups(
@@ -201,6 +130,75 @@ async def list_users(
         )
 
     return user_list
+
+
+@router_users.post("/users/{user_id}/set-groups/", response_model=UserRead)
+async def set_user_groups(
+    user_id: int,
+    user_update: UserUpdateGroups,
+    superuser: UserOAuth = Depends(current_active_superuser),
+    db: AsyncSession = Depends(get_async_db),
+) -> UserRead:
+
+    # Preliminary check that all objects exist in the db
+    user = await _user_or_404(user_id=user_id, db=db)
+    target_group_ids = user_update.group_ids
+    stm = select(func.count(UserGroup.id)).where(
+        UserGroup.id.in_(target_group_ids)
+    )
+    res = await db.execute(stm)
+    count = res.scalar()
+    if count != len(target_group_ids):
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Some UserGroups in {target_group_ids} do not exist.",
+        )
+
+    # Check that default group is not being removed
+    default_group_id = await _get_default_usergroup_id(db=db)
+    if default_group_id not in target_group_ids:
+        raise HTTPException(
+            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
+            detail=(
+                f"Cannot remove user from "
+                f"'{FRACTAL_DEFAULT_GROUP_NAME}' group.",
+            ),
+        )
+
+    # Prepare lists of links to be removed
+    res = await db.execute(
+        select(LinkUserGroup)
+        .where(LinkUserGroup.user_id == user_id)
+        .where(LinkUserGroup.group_id.not_in(target_group_ids))
+    )
+    links_to_remove = res.scalars().all()
+
+    # Prepare lists of links to be added
+    res = await db.execute(
+        select(LinkUserGroup.group_id)
+        .where(LinkUserGroup.user_id == user_id)
+        .where(LinkUserGroup.group_id.in_(target_group_ids))
+    )
+    ids_links_already_in = res.scalars().all()
+    ids_links_to_add = set(target_group_ids) - set(ids_links_already_in)
+
+    # Remove/create links as needed
+    for link in links_to_remove:
+        logger.info(
+            f"Removing LinkUserGroup with {link.user_id=} "
+            f"and {link.group_id=}."
+        )
+        await db.delete(link)
+    for group_id in ids_links_to_add:
+        logger.info(
+            f"Creating new LinkUserGroup with {user_id=} " f"and {group_id=}."
+        )
+        db.add(LinkUserGroup(user_id=user_id, group_id=group_id))
+    await db.commit()
+
+    user_with_groups = await _get_single_user_with_groups(user, db)
+
+    return user_with_groups
 
 
 @router_users.get(