+++ CHANGELOG.md
@@ -16,6 +16,14 @@
     * Add `GET /admin/v2/task-group/activity/` endpoint (\#2005, \#2027).
     * Add `POST /api/v2/task-group/{task_group_id}/{deactivate|reactivate}` endpoints (\#2033, \#2066, \#2078).
     * Add `POST /admin/v2/task-group/{task_group_id}/{deactivate|reactivate}` endpoints (\#2062, \#2078).
+    * Remove `GET /auth/current-user/viewer-paths/` (\#2096).
+    * Add `GET /auth/current-user/allowed-viewer-paths/`, with logic for `fractal-vizarr-viewer` authorization (\#2096).
+    * Add `category`, `modality` and `author` query parameters to `GET /admin/v2/task/` (\#2102).
+    * Add `POST /auth/group/{group_id}/add-user/{user_id}/` (\#2101).
+    * Add `POST /auth/group/{group_id}/remove-user/{user_id}/` (\#2101).
+    * Add `POST /auth/users/{user_id}/set-groups/` (\#2106).
+    * Remove `new_user_ids` property from `PATCH /auth/group/{group_id}/` (\#2101).
+    * Remove `new_group_ids` property from `PATCH /auth/users/{user_id}/` (\#2106).
     * Internals:
       * Fix bug in `_get_collection_task_group_activity_status_message` (\#2047).
       * Remove `valutc` validator for timestamps from API schemas, since it does not match with `psycopg3` behavior (\#2064).
@@ -42,6 +50,7 @@
     * Improve handling of task groups created before 2.9.0 (\#2050).
     * Add `TaskGroupCreateV2Strict` for task collections (\#2080).
     * Always create `script_dir_remote` in SSH lifecycle background tasks (\#2089).
+    * Postpone setting `active=False` in task-group deactivation to after all preliminary checks (\#2100).
 * Runner:
     * Improve error handling in `_zip_folder_to_file_and_remove` (\#2057).
     * Improve error handling in `FractalSlurmSSHExecutor` `handshake` method (\#2083).
@@ -52,6 +61,8 @@
     * Try re-opening the connection in `FractalSSH.check_connection` when an error occurs (\#2035).
     * Move `NoValidConnectionError` exception handling into `FractalSSH.log_and_raise` method (\#2070).
     * Improve closed-socket testing (\#2076).
+* App:
+   * Add `FRACTAL_VIEWER_AUTHORIZATION_SCHEME` and `FRACTAL_VIEWER_BASE_FOLDER` configuration variables (\#2096).
 * Testing:
     * Drop `fetch-depth` from `checkout` in GitHub actions (\#2039).
 * Scripts: